base-requirements.txt
Flask==0.12
Flask-Caching==1.2.0
Flask-Compress==1.0.2
Flask-Cors==3.0.2
Flask-Script==2.0.5
Flask-SSLify==0.1.5
ebaysdk==2.1.4
requests==2.13.0
PyYAML==3.12
docutils==0.13.1
sphinxcontrib-napoleon==0.6.1
python-dateutil==2.6.0
pkutils>=0.13.6,<0.14.0
meza>=0.37.0,<0.38.0
pygogo>=0.9.0,<0.10.0
